Move interfaces Add tests for PipelineBuilder
