+++ src/PipelineFramework.Core/Builder/Interfaces/IComponentResolverHolder.cs
@@ -1,6 +1,6 @@
 ﻿using PipelineFramework.Abstractions;
 
-namespace PipelineFramework.Builder
+namespace PipelineFramework.Builder.Interfaces
 {
     public interface IComponentResolverHolder<out TPipeline, TPayload> where TPipeline : IPipeline
     {
