+++ src/PipelineFramework.Core/Builder/Interfaces/IPipelineBuilder.cs
@@ -1,6 +1,6 @@
 ﻿using PipelineFramework.Abstractions;
 
-namespace PipelineFramework.Builder
+namespace PipelineFramework.Builder.Interfaces
 {
     public interface IPipelineBuilder<out TPipeline, TPayload> where TPipeline : IPipeline
     {
