+++ src/PipelineFramework.Core/Builder/Interfaces/IPipelineComponentHolder.cs
@@ -1,6 +1,6 @@
 ﻿using PipelineFramework.Abstractions;
 
-namespace PipelineFramework.Builder
+namespace PipelineFramework.Builder.Interfaces
 {
     public interface IPipelineComponentHolder<out TPipeline, TComponentBase, TPayload> where TComponentBase : IPipelineComponent where TPipeline : IPipeline
     {
