+++ src/PipelineFramework.Core/Builder/Interfaces/IPipelineComponentNameHolderOrDone.cs
@@ -1,6 +1,6 @@
 ﻿using PipelineFramework.Abstractions;
 
-namespace PipelineFramework.Builder
+namespace PipelineFramework.Builder.Interfaces
 {
     public interface IPipelineComponentNameHolderOrDone<out TPipeline, TPayload> :
         IPipelineComponentNameHolder<TPipeline, TPayload>,
