+++ src/PipelineFramework.Core/Builder/Interfaces/ISettingsHolder.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using PipelineFramework.Abstractions;
 
-namespace PipelineFramework.Builder
+namespace PipelineFramework.Builder.Interfaces
 {
     public interface ISettingsHolder<out TPipeline, TPayload> where TPipeline : IPipeline
     {
